fix(domain): framed_hash cross-platform u64 length prefix (v1.169.140)
PROBLEM (from session-10 addendum 16, deferred at v1.168.139):

crates/sddk-domain/src/identity.rs:412 'fn framed_hash' uses
'domain.len().to_be_bytes()' (usize). On the project's 64-bit
deployment matrix (x86_64/aarch64-linux-musl, x86_64/aarch64-darwin
per .github/workflows/release.yml) usize == u64 == 8 bytes, so the
inline test copies in cli_pack_e2e.rs / cli_approval_e2e.rs /
cli_approval_loop_e2e.rs / ledger_watch.rs — all of which use
'(domain.len() as u64).to_be_bytes()' — coincidentally agree. But
the contract is brittle: if the deployment matrix ever expands to
a 32-bit target (e.g. embedded ARM installer), production would
emit 4-byte length prefixes while tests validate 8-byte prefixes.

Tests-as-source-of-truth-via-coincidence is a known fragile pattern.

FIX: cast domain.len() and part.len() to u64 in production. After
this commit, the framed_hash contract is:

  hasher.update((len as u64).to_be_bytes());    // always 8 bytes
  hasher.update(bytes);

This matches the 4 inline test copies verbatim, so the test
authority no longer disagrees with production by accident; they
genuinely validate the same encoding.

Verified empirically:
  cargo test -p sddk-domain --lib identity
    → 36 passed; 0 failed; 0 ignored (full identity suite GREEN)
  cargo test -p sddk-cli --test cli_pack_e2e
    → 9 passed; 0 failed
  cargo test -p sddk-cli --test cli_approval_e2e
    → 5 passed; 0 failed
  cargo test -p sddk-cli --test ledger_watch
    → 4 passed; 0 failed
  cargo fmt -p sddk-domain -- --check → exit 0 (clean)
  cargo build -p sddk-cli → exit 0

The other 5 files with 'fn fallback_project_id' (cli_explore_e2e,
cli_fork_e2e, cli_stale_e2e, cli_graph_e2e, cli.rs) call
'sddk_domain::resolve_project_identity(...)' directly and have
NEVER drifted. Only the 4 inline-hash copies were affected; this
commit makes production consistent with their encoding choice.

NOT in this commit (deferred, separate concerns):
  - Deduplicating framed_hash: making the production function
    'pub' and importing it from tests would collapse the 5 test
    copies into 1 import. That's a refactor that touches 5 files
    and ~25 lines; per AGENTS.md §2.1 single-concern-per-commit,
    it deserves its own commit. This commit ONLY closes the
    contract drift; the deduplication is logged as a separate
    debt item below.
  - backlog_store::tests::open_owned_* tempdir cleanup (C3bis
    flake from session-10 addendum 11): separate concern.
  - Any other drift between production and tests beyond the
    .len().to_be_bytes() point.

Refs: HANDOFF-2026-09-21-session-10.md addendum 16 (the
characterization was correct, fix was simply not applied). Adds
no new debt; closes nothing pre-existing (the drift was latent,
not yet exploitable).

+++ crates/sddk-domain/src/identity.rs
@@ -412,10 +412,19 @@ pub fn stable_workspace_id(project: &ProjectId, canonical_path: &str) -> String 
 fn framed_hash(domain: &str, parts: &[&str]) -> String {
     use sha2::{Digest, Sha256};
     let mut hasher = Sha256::new();
-    hasher.update(domain.len().to_be_bytes());
+    // Use u64 (not usize) for the length-prefix encoding so that the
+    // hash output is platform-independent. On 64-bit (the project's
+    // deployment matrix: x86_64/aarch64-linux-musl, x86_64/aarch64-darwin
+    // per .github/workflows/release.yml), usize==u64 and both forms
+    // produce the same bytes. On a hypothetical 32-bit target, usize
+    // would be 4 bytes but u64 is always 8 — agreeing with the inline
+    // test copies in cli_approval_loop_e2e.rs / cli_pack_e2e.rs /
+    // cli_approval_e2e.rs / ledger_watch.rs. See addendum 16 of
+    // HANDOFF-2026-09-21-session-10.md for full rationale.
+    hasher.update((domain.len() as u64).to_be_bytes());
     hasher.update(domain.as_bytes());
     for part in parts {
-        hasher.update(part.len().to_be_bytes());
+        hasher.update((part.len() as u64).to_be_bytes());
         hasher.update(part.as_bytes());
     }
     let hash = hasher.finalize();
